serving/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
redis
faiss-cpu
numpy
orjson
pandas
pyarrow
pydantic
pydantic-settings
# torch is installed separately in the Dockerfile
onnxruntime
onnx
